+++ TrafficLightController.h
@@ -17,7 +17,7 @@ class Lamp{
   void turnOn(float seconds);       // Turn On Lamp for 'seconds' seconds
   bool getStatus();                 // Returns 1 for Lamp is on or 0 if off
   void flip();                      // If Lamp is Off, Turn it On & Vice Versa
-  void flash(int timeOn, int timeOff, int numRepeats); // Flash light
+  void flash(float timeOn, float timeOff, int numRepeats); // Flash light
 private:
   int _pinNumber;                   // Stores the pinNumber
   bool _status;                     // Stores whether the light is on 1 or off 0
